Move to array, fix types
