+++ package.json
@@ -71,6 +71,6 @@
     "ts-jest": "^26.5.1",
     "ts-loader": "^8.0.17",
     "ts-node": "^9.1.1",
-    "typescript": "^4.1.3"
+    "typescript": "^4.8.2"
   }
 }
